+++ _studio/shared/umc/codec/h264_dec/include/umc_h264_slice_decoding.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2003-2018 Intel Corporation
+// Copyright (c) 2003-2019 Intel Corporation
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -227,7 +227,7 @@ public:  // DEBUG !!!! should remove dependence
     bool m_isInitialized;
 
 #ifdef MFX_ENABLE_CPLIB
-    uint16_t m_CENCStatusReportNumber;
+    uint32_t m_CENCStatusReportNumber;
 #elif !defined(MFX_PROTECTED_FEATURE_DISABLE)
     uint16_t m_WidevineStatusReportNumber;
 #endif
