+++ _studio/shared/umc/codec/h265_dec/include/umc_h265_cenc_decrypter.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2003-2018 Intel Corporation
+// Copyright (c) 2003-2019 Intel Corporation
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -49,8 +49,8 @@ public:
     UMC::Status GetSliceHeaderPart1(H265SliceHeader * sliceHdr);
     UMC::Status GetSliceHeaderFull(H265CENCSlice *, const H265SeqParamSet *, const H265PicParamSet *);
 
-    uint16_t GetStatusReportNumber() const {return m_CENCStatusReportNumber;}
-    void SetStatusReportNumber(uint16_t statusReportNumber) {m_CENCStatusReportNumber = statusReportNumber;}
+    uint32_t GetStatusReportNumber() const {return m_CENCStatusReportNumber;}
+    void SetStatusReportNumber(uint32_t statusReportNumber) {m_CENCStatusReportNumber = statusReportNumber;}
 
     double GetTime() const {return m_pts;}
     void SetTime(double pts) {m_pts = pts;}
@@ -65,7 +65,7 @@ private:
 
 private:
     double m_pts;
-    uint16_t m_CENCStatusReportNumber;
+    uint32_t m_CENCStatusReportNumber;
 };
 
 #if !defined(OPEN_SOURCE)
@@ -128,7 +128,7 @@ public:
 protected:
     UMC::VideoAccelerator *m_va;
     bool m_bitstreamSubmitted;
-    uint16_t m_PESPacketCounter;
+    uint32_t m_PESPacketCounter;
 
     VADisplay m_dpy;
     VACencStatusBuf m_paramsSet;
